Art color generator
This will generate random colors but it has a problem of preferring cool or neutral colors over warm ones. I plan to fix that! I think it has to do with cool and warm being the topics of interest and then I specified "neutral" for a feeling in the dataset....